internal/gitrepo: extract git functions from pkg/project
